+++ yave/framegraph/FrameGraphPassBuilder.h
@@ -43,7 +43,7 @@ class FrameGraphPassBuilder {
         using render_func = std::function<void(CmdBufferRecorder&, const FrameGraphPass*)>;
 
         FrameGraphMutableImageId declare_image(ImageFormat format, const math::Vec2ui& size);
-        FrameGraphMutableBufferId declare_buffer(usize byte_size);
+        FrameGraphMutableBufferId declare_buffer(u64 byte_size);
 
         FrameGraphMutableImageId declare_copy(FrameGraphImageId src);
         //FrameGraphMutableBufferId declare_copy(FrameGraphBufferId src);
